скрипты 23.03

+++ Assets/Scripts/CharacterControl/HealthVisual.cs
@@ -3,13 +3,13 @@ using UnityEngine;
 public class HealthVisual : MonoBehaviour
 {
     [SerializeField] private HealthManager _healthManager;
-    private int _currentHealth;
+    private float _currentHealth;
     private void Start()
     {
         _healthManager.Add += Visuals;
     }
 
-    private void Visuals(int health)
+    private void Visuals(float health)
     {
         _currentHealth = health;
         Debug.Log($"Current health: {_currentHealth}");
